pyproject.toml
[project]
name = "imager"
version = "0.1.0"
description = ""
authors = [
    {name = "fer",email = "fer@ciandt.com"}
]
readme = "README.md"
requires-python = "~3.12"
package-mode = false
dependencies = [
    "openai (>=1.93.0,<2.0.0)",
    "pillow (>=11.3.0,<12.0.0)",
    "boto3 (>=1.35.0,<2.0.0)",
    "flow-coder @ https://storage.googleapis.com/flow-coder/flow_coder-1.4.0-py3-none-any.whl"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
